pyproject.toml
[project]
name = "fastxai"
version = "0.1.0"
description = "An multi-agents project with langchain(langgraph)1.0，realtime knowledge graph(graphiti-core), vector dbs and LLM..."
readme = "README.md"
requires-python = ">=3.11"
authors = [{name="chenjunfeng", email="cjf_junfeng@163.com"}]

classifiers = [
    "Operating System :: OS Independent",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Framework :: FastAPI",
    "Framework :: FastMCP",
    "Framework :: langchain & langgraph",
    "Framework :: multi-agents",
    "Framework :: realtime knowledge graph",
    "Framework :: graphiti_core",
]

dependencies = [
    "asyncpg==0.31.0",
    "fastapi-structlog==0.7.0",
    "fastapi-users==15.0.1",
    "fastapi[standard]==0.123.9",
    "fastmcp==2.13.3",
    "hypercorn==0.18.0",
    "msgpack==1.1.2",
    "numba==0.62.1",
    "orjson==3.11.4",
    "passlib[bcrypt]==1.7.4",
    "polars==1.35.2",
    "prometheus-fastapi-instrumentator==7.1.0",
    "psutil==7.1.3",
    "psycopg[binary,pool]==3.3.1",
    "psycopg[c]==3.3.1; sys_platform != 'win32'",
    "pydantic-settings==2.12.0",
    "python-jose[cryptography]==3.5.0",
    "sherlock==0.4.1",
    "sqlalchemy[asyncio]==2.0.44",
    "uvicorn[standard]==0.38.0",
    "uvloop>=0.22.1; sys_platform != 'win32'",
    "graphiti-core==0.24.1",
    "langchain==1.1.2",
    "langchain-community>=0.4.1",
    "langchain-openai==1.1.0",
    "langchain-qdrant==1.1.0",
    "langchain-milvus>=0.3.1",
    "langchain-neo4j>=0.6.0",
    "langchain-redis>=0.2.5",
    "langgraph==1.0.4",
    "langgraph-checkpoint-postgres>=3.0.1",
    "duckdb==1.4.2",
    "tenacity>=9.1.2",
    "mysqlclient==2.2.7",
    "asyncmy==0.2.10",
    "celery==5.5.3",
    "flower==2.0.1",
    "redis==6.4.0",
    "fastapi-bgtasks-dashboard>=0.1.5",
    "rq-dashboard-fast>=0.5.12",
    "pydantic-extra-types==2.10.6",
    "bleach>=6.3.0",
    "slowapi>=0.1.9",
    "motor==3.7.1",
    "fastapi-csrf-protect>=1.0.7",
]

[tool.uv]
#index-url = "https://pypi.tuna.tsinghua.edu.cn/simple"
index-url = "https://mirrors.aliyun.com/pypi/simple"

[dependency-groups]
dev = [
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "pytest-mock>=3.15.1",
    "ragas>=0.4.0",
]
